WIP: descriptor: Implement wallet-policies BIP
Policies are simply a restriction on the general variable substitution
mechanism that wally already supports; (a) the variables can only be of
the form @n where n is an integer, and (b) the substitution data must be
an xpub.

See https://github.com/bitcoin/bips/pull/1389

WIP: Requires tests and documentation (and the BIP to be accepted).